load program abstracted
